修改RETROFIT的使用例子 在Infromation中

+++ app/src/main/java/com/example/jungle/weixin/Bean/BaseBean/Remind.java
@@ -6,15 +6,15 @@ package com.example.jungle.weixin.Bean.BaseBean;
 
 public class Remind {
 
-    private int status;
-    private int follower;
-    private int cmt;
-    private int dm;
-    private int mention_status;
-    private int mention_cmt;
-    private int notice;
+    private long status;
+    private long follower;
+    private long cmt;
+    private long dm;
+    private long mention_status;
+    private long mention_cmt;
+    private long notice;
 
-    public Remind(int status, int follower, int cmt, int dm, int mention_status, int mention_cmt, int notice) {
+    public Remind(long status, long follower, long cmt, long dm, long mention_status, long mention_cmt, long notice) {
         this.status = status;
         this.follower = follower;
         this.cmt = cmt;
@@ -24,59 +24,59 @@ public class Remind {
         this.notice = notice;
     }
 
-    public int getStatus() {
+    public long getStatus() {
         return status;
     }
 
-    public void setStatus(int status) {
+    public void setStatus(long status) {
         this.status = status;
     }
 
-    public int getFollower() {
+    public long getFollower() {
         return follower;
     }
 
-    public void setFollower(int follower) {
+    public void setFollower(long follower) {
         this.follower = follower;
     }
 
-    public int getCmt() {
+    public long getCmt() {
         return cmt;
     }
 
-    public void setCmt(int cmt) {
+    public void setCmt(long cmt) {
         this.cmt = cmt;
     }
 
-    public int getDm() {
+    public long getDm() {
         return dm;
     }
 
-    public void setDm(int dm) {
+    public void setDm(long dm) {
         this.dm = dm;
     }
 
-    public int getMention_status() {
+    public long getMention_status() {
         return mention_status;
     }
 
-    public void setMention_status(int mention_status) {
+    public void setMention_status(long mention_status) {
         this.mention_status = mention_status;
     }
 
-    public int getMention_cmt() {
+    public long getMention_cmt() {
         return mention_cmt;
     }
 
-    public void setMention_cmt(int mention_cmt) {
+    public void setMention_cmt(long mention_cmt) {
         this.mention_cmt = mention_cmt;
     }
 
-    public int getNotice() {
+    public long getNotice() {
         return notice;
     }
 
-    public void setNotice(int notice) {
+    public void setNotice(long notice) {
         this.notice = notice;
     }
 }
