+++ app/src/main/java/com/example/jungle/weixin/Bean/BaseBean/User.java
@@ -6,7 +6,7 @@ package com.example.jungle.weixin.Bean.BaseBean;
 
 public class User {
 
-    private int id;
+    private long id;
     private String screen_name;
     private String location;
     private String description;
@@ -26,7 +26,7 @@ public class User {
     private boolean follow_me;
     private Status status;
 
-    public User(int id, String screen_name, String location, String description, String verified_reason, String profile_image_url, String avatar_large, String avatar_hd, String gender, String profile_url, int followers_count, int friends_count, int status_count, boolean following, boolean allow_all_act_msg, boolean allow_all_comment, boolean verified, boolean follow_me, Status status) {
+    public User(long id, String screen_name, String location, String description, String verified_reason, String profile_image_url, String avatar_large, String avatar_hd, String gender, String profile_url, int followers_count, int friends_count, int status_count, boolean following, boolean allow_all_act_msg, boolean allow_all_comment, boolean verified, boolean follow_me, Status status) {
         this.id = id;
         this.screen_name = screen_name;
         this.location = location;
@@ -48,11 +48,11 @@ public class User {
         this.status = status;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
