+++ app/src/main/java/com/example/jungle/weixin/Bean/ParticularBean/ReplyCommentsData.java
@@ -9,8 +9,8 @@ import com.example.jungle.weixin.Bean.BaseBean.User;
  */
 
 public class ReplyCommentsData {
-    private int id;
-    private int mid;
+    private long id;
+    private long mid;
     private String created_at;
     private String text;
     private String source;
